+++ spec/spec_helper.rb
@@ -15,6 +15,8 @@ RSpec.configure do |config|
 
   config.include Helpers
 
+  FactoryGirl.find_definitions
+
   # ## Mock Framework
   #
   # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line: