+++ lib/netzke/basepack/data_adapters/active_record_adapter.rb
@@ -54,7 +54,7 @@ module Netzke::Basepack::DataAdapters
       end
 
       # apply sorting if needed
-      if params[:sort] && sort_params = params[:sort].first
+      if params[:sorters] && sort_params = params[:sorters].first
         dir = sort_params["direction"].downcase
         column = columns.detect { |c| c[:name] == sort_params["property"] }
         relation = apply_sorting(relation, column, dir)