+++ CHANGELOG.md
@@ -6,7 +6,10 @@
   * datetime field in forms respects `read_only` now
   * Re-enable piggybacking commands to the client from within Grid's `get_data` endpoint
 * improvements
-  * Raise a meaninful exception when a model doesn't have a primary key, which is required by Grid
+  * Raise a meaningful exception when a model doesn't have a primary key, which is required by Grid
+  * grid operations are now routed cleanly through endpoints
+* internal changes
+  * `post_data` endpoint in Grid is gone, in favor of `server_create` and `server_update`
 
 # 0.8.3 - 2013-04-24
 * improvements